requirements.txt
h11~=0.14.0
pip~=23.2.1
tornado~=6.4.2
PyYAML~=6.0.2
sniffio~=1.3.1
typing_extensions~=4.12.2
certifi~=2025.1.31
fastapi~=0.115.8
starlette~=0.45.3
pydantic~=2.10.6
pydantic_core~=2.27.2
requests~=2.32.3
anyio~=4.8.0
pywebio~=1.8.3
user-agents~=2.2.0
uvicorn~=0.34.0
tzlocal~=5.3
urllib3~=2.3.0
click~=8.1.8
annotated-types~=0.7.0
idna~=3.10
charset-normalizer~=3.4.1
watchdog~=6.0.0
APScheduler~=3.11.0
ua-parser~=1.0.1
websockets~=15.0
pathspec~=0.12.1